Extract ChatPanel to a dedicated file
