+++ src/chat/chat_history.rs
@@ -1,8 +1,6 @@
-use crate::data::{chats::chat::ChatID, store::Store};
-use chrono::{DateTime, Local, NaiveDateTime, TimeZone};
-
 use makepad_widgets::*;
-use moxin_protocol::data::File;
+use crate::data::store::Store;
+use super::chat_history_card::ChatHistoryCardWidgetRefExt;
 
 live_design! {
     import makepad_widgets::base::*;
@@ -13,69 +11,11 @@ live_design! {
     import makepad_draw::shader::std::*;
 
     import crate::chat::chat_panel::ChatAgentAvatar;
+    import crate::chat::chat_history_card::ChatHistoryCard;
 
     ICON_NEW_CHAT = dep("crate://self/resources/icons/new_chat.svg")
     ICON_CLOSE_PANEL = dep("crate://self/resources/icons/close_left_panel.svg")
     ICON_OPEN_PANEL = dep("crate://self/resources/icons/open_left_panel.svg")
-
-    ChatCard = {{ChatCard}} {
-        content = <RoundedView> {
-            flow: Down
-            width: Fill
-            height: Fit
-            padding: 20
-            spacing: 12
-
-            cursor: Hand
-
-            draw_bg: {
-                color: #fff
-                border_width: 1
-            }
-
-            title = <Label> {
-                width: Fit,
-                height: Fit,
-                draw_text:{
-                    text_style: <BOLD_FONT>{font_size: 10},
-                    color: #000,
-                }
-                text: ""
-            }
-
-            <View> {
-                width: Fill
-                height: Fit
-                align: {y: 1}
-
-                avatar = <ChatAgentAvatar> {
-                    width: 30
-                    height: 30
-
-                    draw_bg: {
-                        radius: 8
-                    }
-                    avatar_label = {
-                        text: ""
-                    }
-                }
-
-                filler = <View> {width: Fill}
-
-                date = <Label> {
-                    width: Fit,
-                    height: Fit,
-                    draw_text:{
-                        text_style: <REGULAR_FONT>{font_size: 10},
-                        color: #667085,
-                    }
-                    text: "5:29 PM, 5/12/24"
-                }
-            }
-
-
-        }
-    }
 
     ChatHistoryActions = <View> {
         spacing: 10,
@@ -115,7 +55,7 @@ live_design! {
         <ChatHistoryActions> {}
 
         list = <PortalList> {
-            ChatCard = <ChatCard> {margin: {top: 20}}
+            ChatHistoryCard = <ChatHistoryCard> {margin: {top: 20}}
         }
     }
 }
@@ -152,9 +92,9 @@ impl Widget for ChatHistory {
                 while let Some(item_id) = list.next_visible_item(cx) {
                     if item_id < chats_count {
                         let mut item = list
-                            .item(cx, item_id, live_id!(ChatCard))
+                            .item(cx, item_id, live_id!(ChatHistoryCard))
                             .unwrap()
-                            .as_chat_card();
+                            .as_chat_history_card();
                         let _ = item.set_chat_id(saved_chat_ids[item_id]);
                         item.draw_all(cx, scope);
                     }
@@ -169,132 +109,10 @@ impl Widget for ChatHistory {
 impl WidgetMatchEvent for ChatHistory {
     fn handle_actions(&mut self, cx: &mut Cx, actions: &Actions, scope: &mut Scope) {
         let store = scope.data.get_mut::<Store>().unwrap();
-        let widget_uid = self.widget_uid();
 
         if self.button(id!(new_chat_button)).clicked(&actions) {
             store.chats.create_empty_chat();
             self.redraw(cx);
-            cx.widget_action(widget_uid, &scope.path, ChatHistoryAction::NewChat);
         }
     }
-}
-
-#[derive(Live, LiveHook, Widget)]
-pub struct ChatCard {
-    #[deref]
-    view: View,
-    #[rust]
-    chat_id: ChatID,
-}
-
-impl Widget for ChatCard {
-    fn handle_event(&mut self, cx: &mut Cx, event: &Event, scope: &mut Scope) {
-        self.view.handle_event(cx, event, scope);
-        self.widget_match_event(cx, event, scope);
-    }
-
-    fn draw_walk(&mut self, cx: &mut Cx2d, scope: &mut Scope, walk: Walk) -> DrawStep {
-        let store = scope.data.get_mut::<Store>().unwrap();
-        let chat = store
-            .chats
-            .saved_chats
-            .iter()
-            .find(|c| c.borrow().id == self.chat_id)
-            .unwrap();
-
-        if let Some(current_chat_id) = store.chats.get_current_chat_id() {
-            let content_view = self.view(id!(content));
-
-            if current_chat_id == self.chat_id {
-                let active_border_color = vec3(0.082, 0.522, 0.604);
-                content_view.apply_over(
-                    cx,
-                    live! {
-                        draw_bg: {border_color: (active_border_color)}
-                    },
-                );
-            } else {
-                let border_color = vec3(0.918, 0.925, 0.941);
-                content_view.apply_over(
-                    cx,
-                    live! {
-                        draw_bg: {border_color: (border_color)}
-                    },
-                );
-            }
-        }
-
-        let title_label = self.view.label(id!(title));
-        title_label.set_text(chat.borrow_mut().get_title());
-
-        let initial_letter = chat
-            .borrow()
-            .model_filename
-            .chars()
-            .next()
-            .unwrap_or_default()
-            .to_uppercase()
-            .to_string();
-
-        let avatar_label = self.view.label(id!(avatar.avatar_label));
-        avatar_label.set_text(&initial_letter);
-
-        let date_label = self.view.label(id!(date));
-
-        // Format date.
-        // TODO: Feels wrong to asume the id will always be the date, do smth about this.
-        let naive_datetime = NaiveDateTime::from_timestamp_millis(chat.borrow().id as i64)
-            .expect("Invalid timestamp");
-        let datetime: DateTime<Local> = Local.from_utc_datetime(&naive_datetime);
-        let formatted_date = datetime.format("%-I:%M %p, %-d/%m/%y").to_string();
-
-        date_label.set_text(&formatted_date);
-
-        self.view.draw_walk(cx, scope, walk)
-    }
-}
-
-impl WidgetMatchEvent for ChatCard {
-    fn handle_actions(&mut self, cx: &mut Cx, actions: &Actions, scope: &mut Scope) {
-        let store = scope.data.get_mut::<Store>().unwrap();
-        let widget_uid = self.widget_uid();
-
-        if let Some(fe) = self.view(id!(content)).finger_down(actions) {
-            if fe.tap_count == 1 {
-                cx.widget_action(
-                    widget_uid,
-                    &scope.path,
-                    ChatHistoryAction::ChatSelected(self.chat_id),
-                );
-
-                store.select_chat(self.chat_id);
-
-                self.redraw(cx);
-            }
-        }
-    }
-}
-
-impl ChatCard {
-    pub fn set_chat_id(&mut self, id: ChatID) {
-        self.chat_id = id;
-    }
-}
-
-impl ChatCardRef {
-    pub fn set_chat_id(&mut self, id: ChatID) -> Result<(), &'static str> {
-        let Some(mut inner) = self.borrow_mut() else {
-            return Err("Widget not found in the document");
-        };
-
-        inner.set_chat_id(id);
-        Ok(())
-    }
-}
-
-#[derive(Clone, DefaultNone, Eq, Hash, PartialEq, Debug)]
-pub enum ChatHistoryAction {
-    None,
-    ChatSelected(ChatID),
-    NewChat,
 }
